Cleaned-up contracts and fixed Groups API

+++ Iggy_SDK/Contracts/Http/MessageResponse.cs
@@ -4,7 +4,7 @@ namespace Iggy_SDK.Contracts;
 
 public sealed class MessageResponse
 {
-	public required int Offset { get; init; }
+	public required ulong Offset { get; init; }
 	public required ulong Timestamp { get; init; }
 	public UInt128 Id { get; init; }
 	public required string Payload { get; init; }
