+++ Iggy_SDK/Contracts/Http/OffsetContract.cs
@@ -4,5 +4,5 @@ public sealed class OffsetContract
 {
 	public required int ConsumerId { get; init; }
 	public required int PartitionId { get; init; }
-	public required int Offset { get; init; }
+	public required ulong Offset { get; init; }
 }
